% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ShannonEntropy.R
\name{compute_ShannonEntropy}
\alias{compute_ShannonEntropy}
\title{Computes Shannon entropy}
\usage{
compute_ShannonEntropy(data = data, exp_thr = 1, cell_specificity_thr = 1)
}
\arguments{
\item{data}{A data frame containing expression data with gene identifier as
first column and normalized expression values in RPKM/FPKM, TPM, CPM or any
in other columns}

\item{exp_thr}{A numeric value to filter gene expression}

\item{cell_specificity_thr}{A numeric value for Shannon entropy threshold to
define cellular specificity}
}
\value{
A data frame containing cell type or condition specific expression
data in the same format as input data
}
\description{
Computes Shannon entropy for gene expression across cell types or conditions
to find cell type or condition specific gene expression patterns
genomic regions.
}
\details{
Reference papers
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4932457/pdf/FEB4-6-774.pdf
https://genomebiology.biomedcentral.com/track/pdf/10.1186/gb-2005-6-4-r33?
site=genomebiology.biomedcentral.com
https://www.nature.com/nature/journal/v534/n7609/pdf/nature18606.pdf
}
\examples{
\dontrun{
cell_type_data <- compute_ShannonEntropy(data = data, exp_thr = 1,
cell_specificity_thr = 1)
}
}
